Extract PhotoProcessingLoggers to leaf file and bypass barrel import
- Create photo/logger-types.ts as a dependency-free leaf defining
  PhotoLogger and PhotoProcessingLoggers interfaces.
- Update logger-adapter.ts to re-export from the leaf, removing the
  duplicate interface definition.
- Change image/histogram.ts to import getGlobalLoggers directly from
  photo/logger-adapter.js instead of the photo/ barrel, eliminating
  the data-processors → histogram → photo/index barrel cycle.

Builder circular deps: 22 → 20.
